requirements.txt
astropy==6.1.3
matplotlib==3.7.2
mpl_scatter_density==0.7
numpy==1.25.0
pandas==2.2.3
Pillow==11.1.0
PyYAML==6.0.2
scikit_learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
torch==2.3.1+cu118
torchsummary==1.5.1
torchvision==0.18.1+cu118
torchviz==0.0.2
umap_learn==0.5.6
wandb==0.19.1
